upstream updates to chips-core from @barrystyle

+++ src/wallet/scriptpubkeyman.cpp
@@ -546,7 +546,7 @@ int64_t LegacyScriptPubKeyMan::GetOldestKeyPoolTime() const
     return oldestKey;
 }
 
-size_t LegacyScriptPubKeyMan::KeypoolCountExternalKeys() const
+uint64_t LegacyScriptPubKeyMan::KeypoolCountExternalKeys() const
 {
     LOCK(cs_KeyStore);
     return setExternalKeyPool.size() + set_pre_split_keypool.size();
